chore: move ErrorLogger and its implementations
