+++ app/src/main/java/com/pereyrarg11/drinks/core/logger/error/ErrorLogger.kt
@@ -1,4 +1,4 @@
-package com.pereyrarg11.drinks.core.util.error
+package com.pereyrarg11.drinks.core.logger.error
 
 
 interface ErrorLogger {
