+++ app/src/main/java/com/pereyrarg11/drinks/core/logger/error/ErrorLoggerMediator.kt
@@ -1,4 +1,4 @@
-package com.pereyrarg11.drinks.core.util.error
+package com.pereyrarg11.drinks.core.logger.error
 
 import javax.inject.Inject
 import javax.inject.Singleton
